net: close review round 5 (ship verdict) - doc/code agreement and one wrong comment
Round 5 returned SHIP: no sixth canonicality hole across ~1.0M decodable
mutants through an oracle written from scratch (three corpora plus a column
oracle), 0 second spellings; all five of round 4's new rows verified
load-bearing by revert-and-run, each flipping exactly its own row. Its
findings are hygiene and one comment that misstated the spec - none touches
the hashed bytes - and all are applied here.

1 (medium) - net_internal.h claimed "the archive refuses more than
MAX_LOG_RECORDS_PER_PACKET records at one effective_tick (§20.2.9, the
bound archive_encode_segment TL_CHECKs)". All three clauses were wrong: the
encoder's TL_CHECK is an AGGREGATE (8 * tick_count), and §20.2.9 states no
per-tick bound at all - 16 records at one tick over a 2-tick span encode and
decode cleanly. The store's per-tick rule is therefore STRICTER THAN THE
FORMAT: sufficient for the aggregate, and it does stop a caller assembling a
set the encoder aborts on, but it refuses records a valid segment could
carry. The comments now say exactly that, and whether the bound belongs in
the format is filed as a ruling request with a recommendation.

2 (low) - the segment-to-file binding was asserted in a comment and
implemented nowhere. It cannot live inside archive_decode_segment, which
never sees the file header, so it is the reader's obligation - now named
(archive_check_segment_file), stated in §20.2.9, and tested.

3 (low) - wire.h still quoted the pre-amendment docs: the §20.2.9 "0..35 /
36 streams per slot" line the ruling replaced with 35 channels, and a note
saying §20.2 "should say so" when it now does. ARCHIVE_CH_COUNT is gone -
it kept a figure the doc no longer states and nothing but a test referenced
it. §20.2's own exemption sentence also claimed all three interior records
carry sizeof/offsetof pins and little-endian discipline; ArchiveStreamHeader
carries neither, five paragraphs above where the same section says its wire
form is two uvarints. Corrected.

4 (low) - two test rows pinned the retired 8-byte stream header and
ARCHIVE_CH_COUNT: layout protection for bytes that no longer cross the wire.
Replaced with pins on what IS on the wire - the key's radix, the last legal
channel, and the bound. This is round 4's own lesson applied to the file
that taught it.

5 (info) - one u32 slot-index expression among u64 ones. Unreachable today
(ERR_NET_CAPACITY fires first) and one edit to MAX_PEERS away from not
being.

6 - round 5 corrected a FILED item of mine: the O(tick_count^2) entry said
the caller's frame capacity was the only bound. out_record_capacity bounds
it too, so the scans are min(8*tick_count, out_record_capacity)^2 and the
61 s figure needs a 320,000-entry record buffer as well. The filing is
amended; the recommendation stands, since both bounds are buffer sizes
rather than format rules.

62/62 net rows green; full fast suite 385 passed / 0 failed; non-dev tiers
clean with -Wconversion -Wsign-conversion; docaudit and includes clean.

+++ src/net/archive.cpp
@@ -397,7 +397,12 @@ ErrCode archive_decode_segment(ByteReader* r, ArchiveSegmentHeader* out_header,
 
     // Slots outside slot_mask decode as ZERO (docs/NETCODE.md §20.2.9).
     for (u32 s = 0; s < MAX_PEERS; ++s) {
-        for (u32 i = 0; i < tick_count; ++i) { out_frames[s * tick_count + i] = wire_zero_frame(); }
+        // (u64) like every other slot index here: the wrap needs tick_count >= 2^32/7 and a
+        // ~373 GB caller buffer, so ERR_NET_CAPACITY fires first today - but that is one edit to
+        // MAX_PEERS or the capacity contract away from being false.
+        for (u32 i = 0; i < tick_count; ++i) {
+            out_frames[(u64)s * tick_count + i] = wire_zero_frame();
+        }
     }
 
     // Empty streams are OMITTED on the wire, so the stream region is read until it ends rather
